Update 20190923202523_create_uri_transitions.rb
Try bigint instead

+++ db/migrate/20190923202523_create_uri_transitions.rb
@@ -4,7 +4,7 @@ class CreateUriTransitions < ActiveRecord::Migration[5.2]
       t.string :to_state, null: false
       t.text :metadata
       t.integer :sort_key, null: false
-      t.integer :solr_document_uri_id, null: false
+      t.bigint :solr_document_uri_id, null: false
       t.boolean :most_recent, null: false
 
       # If you decide not to include an updated timestamp column in your transition
